+++ core/src/main/java/com/shatteredpixel/shatteredpixeldungeon/Rankings.java
@@ -72,7 +72,7 @@ public enum Rankings {
 
 	public Record latestDaily;
 	public Record latestDailyReplay = null; //not stored, only meant to be temp
-	public LinkedHashMap<Long, Integer> dailyScoreHistory = new LinkedHashMap<>();
+	public LinkedHashMap<Long, Long> dailyScoreHistory = new LinkedHashMap<>();
 
 	public void submit( boolean win, Object cause ) {
 
@@ -160,12 +160,12 @@ public enum Rankings {
 		save();
 	}
 
-	private int score( boolean win ) {
+	private long score(boolean win ) {
 		return (Statistics.goldCollected + Dungeon.hero.lvl * (win ? 26 : Dungeon.depth ) * 100) * (win ? 2 : 1);
 	}
 
 	//assumes a ranking is loaded, or game is ending
-	public int calculateScore(){
+	public long calculateScore(){
 
 		if (Dungeon.initialVersion > 453){
 			Statistics.progressScore = Dungeon.hero.lvl * Statistics.deepestFloor * 25;
@@ -374,7 +374,7 @@ public enum Rankings {
 		bundle.put(LATEST_DAILY, latestDaily);
 
 		long[] dates = new long[dailyScoreHistory.size()];
-		int[] scores = new int[dailyScoreHistory.size()];
+		long[] scores = new long[dailyScoreHistory.size()];
 		int i = 0;
 		for (Long l : dailyScoreHistory.keySet()){
 			dates[i] = l;
@@ -426,7 +426,7 @@ public enum Rankings {
 				latestDaily = (Record) bundle.get(LATEST_DAILY);
 
 				dailyScoreHistory.clear();
-				int[] scores = bundle.getIntArray(DAILY_HISTORY_SCORES);
+				long[] scores = bundle.getLongArray(DAILY_HISTORY_SCORES);
 				int i = 0;
 				long latestDate = 0;
 				for (long date : bundle.getLongArray(DAILY_HISTORY_DATES)){
@@ -474,7 +474,7 @@ public enum Rankings {
 		public String gameID;
 
 		//Note this is for summary purposes, visible score should be re-calculated from game data
-		public int score;
+		public long score;
 
 		public String customSeed;
 		public boolean daily;
@@ -511,7 +511,7 @@ public enum Rankings {
 			}
 			
 			win		    = bundle.getBoolean( WIN );
-			score	    = bundle.getInt( SCORE );
+			score	    = bundle.getLong( SCORE );
 			customSeed  = bundle.getString( SEED );
 			daily       = bundle.getBoolean( DAILY );
