+++ core/src/main/java/com/shatteredpixel/shatteredpixeldungeon/items/armor/curses/Bulk.java
@@ -33,7 +33,7 @@ public class Bulk extends Armor.Glyph {
 	private static ItemSprite.Glowing BLACK = new ItemSprite.Glowing( 0x000000 );
 	
 	@Override
-	public int proc(Armor armor, Char attacker, Char defender, int damage) {
+	public long proc(Armor armor, Char attacker, Char defender, long damage) {
 		
 		//no proc effect, see armor.speedfactor
 		return damage;
