+++ core/src/main/java/com/shatteredpixel/shatteredpixeldungeon/items/armor/curses/Displacement.java
@@ -35,7 +35,7 @@ public class Displacement extends Armor.Glyph {
 	private static ItemSprite.Glowing BLACK = new ItemSprite.Glowing( 0x000000 );
 
 	@Override
-	public int proc(Armor armor, Char attacker, Char defender, int damage ) {
+	public long proc(Armor armor, Char attacker, Char defender, long damage ) {
 
 		float procChance = 1/20f * procChanceMultiplier(defender);
 		if ( Random.Float() < procChance ) {
