+++ core/src/main/java/com/shatteredpixel/shatteredpixeldungeon/items/armor/glyphs/Affection.java
@@ -39,9 +39,9 @@ public class Affection extends Glyph {
 	private static ItemSprite.Glowing PINK = new ItemSprite.Glowing( 0xFF4488 );
 	
 	@Override
-	public int proc( Armor armor, Char attacker, Char defender, int damage) {
+	public long proc( Armor armor, Char attacker, Char defender, long damage) {
 
-		int level = Math.max(0, armor.buffedLvl());
+		long level = Math.max(0, armor.buffedLvl());
 		
 		// lvl 0 - 15%
 		// lvl 1 ~ 19%
