+++ core/src/main/java/com/shatteredpixel/shatteredpixeldungeon/items/armor/glyphs/Entanglement.java
@@ -41,9 +41,9 @@ public class Entanglement extends Glyph {
 	private static ItemSprite.Glowing BROWN = new ItemSprite.Glowing( 0x663300 );
 	
 	@Override
-	public int proc(Armor armor, Char attacker, final Char defender, final int damage ) {
+	public long proc(Armor armor, Char attacker, final Char defender, final long damage ) {
 
-		final int level = Math.max( 0, armor.buffedLvl() );
+		final long level = Math.max( 0, armor.buffedLvl() );
 		float procChance = 1/4f * procChanceMultiplier(defender);
 
 		if (Random.Float() < procChance) {
