+++ core/src/main/java/com/shatteredpixel/shatteredpixeldungeon/items/armor/glyphs/Potential.java
@@ -38,9 +38,9 @@ public class Potential extends Glyph {
 	private static ItemSprite.Glowing WHITE = new ItemSprite.Glowing( 0xFFFFFF, 0.6f );
 	
 	@Override
-	public int proc( Armor armor, Char attacker, Char defender, int damage) {
+	public long proc( Armor armor, Char attacker, Char defender, long damage) {
 
-		int level = Math.max( 0, armor.buffedLvl() );
+		long level = Math.max( 0, armor.buffedLvl() );
 		
 		// lvl 0 - 16.7%
 		// lvl 1 - 28.6%
